requirements.txt
numpy
pandas
scikit-learn
matplotlib
seaborn
scipy
jupyter
statsmodels
xgboost
fastapi
uvicorn
python-dotenv